Working multiple powerups display - When collected more than one of bomb count or range powerups, will now display the amount gathered. If reached the max amount allowed, will instead display "max". - Misc changes to Text class in order to accomadate for the new InfoPanel additions. - Moved Text texture setting to Animations instead of InfoPanel to avoid any non-loaded texture issues. - Made game static members private and added getters to access them. Classes only need the data, not to be able to modify them. - Went through the project and tried to remove as many compiler warnings as I could. Still getting fallthrough warnings in Enemy though, as well as a linker library conflict. - New constans of course because why would there not be??

+++ Bomberman/Player.cpp
@@ -11,13 +11,13 @@ using std::cout, std::endl;
 
 Player::Player(const sf::Texture& tex, Pod (&pods)[_rows][_cols],
 	std::vector<Bomb>& bombs, std::vector<Explosion>& explosions) :
-	Entity(tex, pods), bombs(bombs), explosions(explosions),
-	speed(_playerSpeed * _speedScale),
-	joyX(0), joyY(0), lives(3),
-	blast(1), maxBombs(1), wait(0), remote(false),
-	isFireShield(false), isInvincible(false),
-	wallPhase(false), bombPhase(false),
-	justDied(false)
+		Entity(tex, pods), bombs(bombs), explosions(explosions),
+		speed(_playerSpeed * _speedScale),
+		joyX(0), joyY(0), lives(3),
+		blast(1), maxBombs(1), wait(0), remote(false),
+		isFireShield(false), isInvincible(false),
+		wallPhase(false), bombPhase(false),
+		justDied(false)
 {
 	setTexture(64, 0);
 	setPosition(1, 1);
@@ -38,14 +38,14 @@ void Player::update()
 		joyX += (isKeyPressed(Scan::Right) - isKeyPressed(Scan::Left));
 		joyY += (isKeyPressed(Scan::Down) - isKeyPressed(Scan::Up));
 
-		joyX += std::round(sf::Joystick::getAxisPosition(0, sf::Joystick::Axis::X));
-		joyY += std::round(sf::Joystick::getAxisPosition(0, sf::Joystick::Axis::Y));
+		joyX += static_cast<int>(std::round(sf::Joystick::getAxisPosition(0, sf::Joystick::Axis::X)));
+		joyY += static_cast<int>(std::round(sf::Joystick::getAxisPosition(0, sf::Joystick::Axis::Y)));
 
 		moveLogic();
 
 		// Spawn a bomb
 		if (isKeyPressed(Scancode::Z))
-			if (!pods[tileY][tileX].isFilled && bombs.size() < maxBombs && !pods[tileY][tileX].isExit)
+			if (!pods[tileY][tileX].isFilled && static_cast<int>(bombs.size()) < maxBombs && !pods[tileY][tileX].isExit)
 			{
 				pods[tileY][tileX].isFilled = true;
 				pods[tileY][tileX].isHard = true;
